Added fix for #9
It seems that Umbraco may on occasion index a date in another format than it usually does, and if so, this would cause the GetSortValueByDateTime to throw an exception as DateTime.ParseExact would try to parse a value in a different format from what we have specified.

With this commit, we're instead using DateTime.TryParseExact, which doesn't throw an exception if it meets an urecognized format. Instead the GetSortValueByDateTime will now fall back to the result's create date instead - as well as append a warning to Umbraco's log.