Added "logger" parameter to "IPostSortOptions.Sort" method
Since we're often using the static "SearchUtils.Sorting.GetSortValueByDateTime" utility method, which now requires an "ILogger" instance for logging date parsing errors, we need a way to pass the "ILogger" to the method.

I'm not 100% sure yet this is the correct approach. Maybe we shouldn't even have static methods for this. But if so, we can have a look at this later.